requirements.txt
pandas
numpy

langchain
langchain-community
langchain-core
langchain-text-splitters

langchain-chroma
chromadb
sentence-transformers
torch
python-dotenv
tqdm

streamlit
